+++ packages/manager/src/components/Autocomplete/Autocomplete.stories.tsx
@@ -6,7 +6,6 @@ import { styled } from '@mui/material/styles';
 import { action } from '@storybook/addon-actions';
 import React, { useState } from 'react';
 
-import { Country } from 'src/components/RegionSelect/utils';
 import { Flag } from 'src/components/Flag';
 import { IconButton } from 'src/components/IconButton';
 import { List } from 'src/components/List';
@@ -19,6 +18,7 @@ import { SelectedIcon } from './Autocomplete.styles';
 
 import type { EnhancedAutocompleteProps } from './Autocomplete';
 import type { Meta, StoryFn, StoryObj } from '@storybook/react';
+import type { Country } from 'src/components/RegionSelect/RegionSelect.types';
 
 const LABEL = 'Select a Linode';
