+++ packages/manager/src/components/RegionSelect/RegionSelect.constants.ts
@@ -1,5 +1,4 @@
-// Thank you https://github.com/BRIXTOL/country-continent
-
+// See https://github.com/BRIXTOL/country-continent
 export const COUNTRY_CODE_TO_CONTINENT_CODE = Object.freeze({
   AD: 'EU',
   AE: 'AS',
@@ -232,7 +231,7 @@ export const COUNTRY_CODE_TO_CONTINENT_CODE = Object.freeze({
   TZ: 'AF',
   UA: 'EU',
   UG: 'AF',
-  UK: 'EU', // for compatability with the Linode API
+  UK: 'EU', // for compatibility with the Linode API
   UM: 'OC',
   US: 'NA',
   UY: 'SA',
@@ -253,8 +252,6 @@ export const COUNTRY_CODE_TO_CONTINENT_CODE = Object.freeze({
   ZM: 'AF',
   ZW: 'AF',
 });
-
-export type Country = keyof typeof COUNTRY_CODE_TO_CONTINENT_CODE;
 
 export const CONTINENT_CODE_TO_CONTINENT = Object.freeze({
   AF: 'Africa',
