+++ packages/manager/src/components/RegionSelect/RegionSelect.types.ts
@@ -1,0 +1,27 @@
+import { COUNTRY_CODE_TO_CONTINENT_CODE } from './RegionSelect.constants';
+
+export type Country = keyof typeof COUNTRY_CODE_TO_CONTINENT_CODE;
+
+import type { Region } from '@linode/api-v4';
+import type { EnhancedAutocompleteProps } from 'src/components/Autocomplete/Autocomplete';
+
+export interface OptionType {
+  data?: any;
+  label: string;
+  value: string;
+}
+
+export interface RegionSelectProps
+  extends Omit<
+    EnhancedAutocompleteProps<OptionType, false>,
+    'label' | 'onChange' | 'options'
+  > {
+  handleSelection: (id: string) => void;
+  helperText?: string;
+  isClearable?: boolean;
+  label?: string;
+  regions: Region[];
+  required?: boolean;
+  selectedId: null | string;
+  width?: number;
+}
